+++ CHANGELOG.md
@@ -25,6 +25,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.35.18] - 2025-07-03
+
+### Added
+- Support Jinja2 rendering of the `if_exists` property of CFS configurations,
+  IMS images, and BOS session templates defined in the bootprep input file.
+
 ## [3.35.17] - 2025-06-26
 
 ### Security